refactor : NameCard -> PhotoCard로 네이밍 변경
