+++ feature/namecard/src/main/java/com/wepli/feature/namecard/detail/viewmodel/NameCardDetailViewModel.kt
@@ -4,7 +4,7 @@ import base.BaseMviViewModel
 import com.wepli.feature.namecard.detail.mvi.NameCardDetailEffect
 import com.wepli.feature.namecard.detail.mvi.NameCardDetailIntent
 import com.wepli.feature.namecard.detail.mvi.NameCardDetailUiState
-import com.wepli.shared.feature.uimodel.namecard.NameCardUiData
+import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
 import com.wepli.uimodel.music.SongUiData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -104,7 +104,7 @@ class NameCardDetailViewModel @Inject constructor(
         state: NameCardDetailUiState
     ) = withContext(Dispatchers.IO) {
         val user = userRepository.getUser()
-        NameCardUiData(
+        PhotoCardUiData(
             nickname = user?.nickname ?: "",
             profileImg = user?.profileImgUrl ?: "",
             userTendency = "Melody Memories",
