+++ feature/namecard/src/main/java/com/wepli/feature/namecard/main/viewmodel/NameCardMainViewModel.kt
@@ -6,7 +6,7 @@ import com.wepli.feature.namecard.main.mvi.NameCardMainIntent
 import com.wepli.feature.namecard.main.mvi.NameCardMainUiState
 import com.wepli.shared.feature.mock.songMockData
 import com.wepli.shared.feature.mock.userMockData
-import com.wepli.shared.feature.uimodel.namecard.NameCardUiData
+import com.wepli.shared.feature.uimodel.namecard.PhotoCardUiData
 import com.wepli.shared.feature.uimodel.user.UserUiData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import repository.user.UserRepository
@@ -35,7 +35,7 @@ class NameCardMainViewModel @Inject constructor(
 
     private fun setNameCardInfo() {
         val user = userMockData.random()
-        val nameCardInfo = NameCardUiData(
+        val nameCardInfo = PhotoCardUiData(
             nickname = user.nickname,
             userTendency = "Melody Memories",
             profileImg = user.profileImgUrl,
